(agentxmpp) Do not die silently when the config file is empty
When the client agent starts it check for the agentconf.ini file.
It may happen that this file is empty.

The current behaviour is to die silently. This is not the good behaviour
because we cannot know what happens without digging and starting the
agent manually in a cmd.

Now we add 2 new mecanisms.

-> If the config file is empty we catch the error and we add it on the
logs.
-> We add a agentconf.ini.tpl file that we will copy in the same place
of the emply file.
And then we reread the configuraton.