(feature) Add windows updates support.
This feature in pulse add support for Wsus.

Add a new function search_system_info_reg

This function is used to retrieve informations about the version of the
client windows.

We retrieve informations like:
        - CurrentBuild
        - CurrentVersion
        - InstallationType
        - ProductName
        - ReleaseId
        - DisplayVersion
        - RegisteredOwner

With the function search_history_update() we are able to list all the
already instaled KBs on the client machine.